Add files via upload
A dinosaur code for being more proactive in contacting clients. Filters a client list and sends an automated email containing a list of clients that a due for contact. There are two types of contacts; a full review and a reach out. Clients are categorized by A, B, C and D for the frequency (annually, semiannually, quarter-annually) and type of contact. Meant to be triggered by a Outlook reminder at the beginning of every month.